+++ CdIts.Caffoa.Tests/TestResultHandler.cs
@@ -13,6 +13,6 @@ public class TestResultHandler
     {
         var test = new SubData();
         var handler = new CaffoaEarlySerializingResultHandler();
-        handler.Invoking(h => h.Json(test, 200)).Should().Throw<JsonSerializationException>();
+        handler.Invoking(h => h.Result(test, 200, [])).Should().Throw<JsonSerializationException>();
     }
 }