+++ demo/DemoV2/DemoV2UserFunctions.generated.cs
@@ -50,7 +50,7 @@ namespace DemoV2
                     limitValue = _converter.Parse<int>(limitQueryValue, "limit");
                 await using var instance = _factory.Instance(request);
                 var result = await instance.UsersGetAsync(offsetValue, limitValue, request.HttpContext.RequestAborted);
-                return _resultHandler.Json(result, 200);
+                return _resultHandler.Result(result, 200, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {
@@ -78,7 +78,7 @@ namespace DemoV2
                     _ => throw _errorHandler.WrongContent("type", discriminator, new [] { "simple", "guest" })
                 };
                 var result = await task;
-                return _resultHandler.Json(result, 201);
+                return _resultHandler.Result(result, 201, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {
@@ -106,7 +106,7 @@ namespace DemoV2
                     _ => throw _errorHandler.WrongContent("type", discriminator, new [] { "simple", "guest" })
                 };
                 var (result, code) = await task;
-                return _resultHandler.Json(result, code);
+                return _resultHandler.Result(result, code, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {
@@ -126,7 +126,7 @@ namespace DemoV2
             try {
                 await using var instance = _factory.Instance(request);
                 var result = await instance.UserPatchAsync(userId, await _jsonParser.Parse<JObject>(request.Body), request.HttpContext.RequestAborted);
-                return _resultHandler.Json(result, 200);
+                return _resultHandler.Result(result, 200, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {
@@ -146,7 +146,7 @@ namespace DemoV2
             try {
                 await using var instance = _factory.Instance(request);
                 var result = await instance.UserGetAsync(userId, request.HttpContext.RequestAborted);
-                return _resultHandler.Json(result, 200);
+                return _resultHandler.Result(result, 200, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {
@@ -206,7 +206,7 @@ namespace DemoV2
             try {
                 await using var instance = _factory.Instance(request);
                 var result = await instance.UsersGetByBirthdateAsync(_converter.ParseDateOnly(date, "date"), request.HttpContext.RequestAborted);
-                return _resultHandler.Json(result, 200);
+                return _resultHandler.Result(result, 200, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {
@@ -239,7 +239,7 @@ namespace DemoV2
                     maxResultsValue = _converter.Parse<int>(maxResultsQueryValue, "maxResults");
                 await using var instance = _factory.Instance(request);
                 var result = await instance.UsersSearchByDateAsync(beforeValue, afterValue, maxResultsValue, request.HttpContext.RequestAborted);
-                return _resultHandler.Json(result, 200);
+                return _resultHandler.Result(result, 200, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {
@@ -259,7 +259,7 @@ namespace DemoV2
             try {
                 await using var instance = _factory.Instance(request);
                 var result = await instance.GetTagsAsync(request.HttpContext.RequestAborted);
-                return _resultHandler.Json(result, 200);
+                return _resultHandler.Result(result, 200, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {
@@ -279,7 +279,7 @@ namespace DemoV2
             try {
                 await using var instance = _factory.Instance(request);
                 var result = await instance.GetUserTagsAsync(request.HttpContext.RequestAborted);
-                return _resultHandler.Json(result, 200);
+                return _resultHandler.Result(result, 200, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {
@@ -311,7 +311,7 @@ namespace DemoV2
                     excludeValue = _converter.ParseEnumArray<MyEnumType>(_jsonParser, excludeQueryValue, "exclude");
                 await using var instance = _factory.Instance(request);
                 var result = await instance.ListEnumsAsync(filterValue, includeValue, flagsValue, excludeValue, request.HttpContext.RequestAborted);
-                return _resultHandler.Json(result, 200);
+                return _resultHandler.Result(result, 200, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {
@@ -331,7 +331,7 @@ namespace DemoV2
             try {
                 await using var instance = _factory.Instance(request);
                 var result = await instance.ListEnums2Async(_converter.ParseEnum<MyEnumType>(filter, "filter"), request.HttpContext.RequestAborted);
-                return _resultHandler.Json(result, 200);
+                return _resultHandler.Result(result, 200, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {
@@ -351,7 +351,7 @@ namespace DemoV2
             try {
                 await using var instance = _factory.Instance(request);
                 var result = await instance.EchoOneOfAsync(await _jsonParser.Parse<GroupedOneOf>(request.Body), request.HttpContext.RequestAborted);
-                return _resultHandler.Json(result, 200);
+                return _resultHandler.Result(result, 200, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {
@@ -371,7 +371,7 @@ namespace DemoV2
             try {
                 await using var instance = _factory.Instance(request);
                 var result = await instance.EchoOneOfArrayAsync(await _jsonParser.Parse<IEnumerable<AnyUser>>(request.Body), request.HttpContext.RequestAborted);
-                return _resultHandler.Json(result, 200);
+                return _resultHandler.Result(result, 200, request.Headers?.Accept ??  Array.Empty<string>());
             } catch(CaffoaClientError err) {
                 return err.Result;
             } catch (Exception e) {