+++ changelog.md
@@ -2,6 +2,7 @@
 ## 5.0.0
 * Support for net 10.0 and net9.0
 * Dropped support for netcore 3.1 and net7.0
+* Renamed interface method `Json` to `Result` in `ICaffoaResultHandler`, and added accept header data. This can be used to return different content data than json.
 * Dropped support for System.Text.Json net6.0 (allow to drop legacy support for missing feature that where introduced in net7.0)
 * Drop support for `flavor: SystemTextJsonPre7` (use `flavor: SystemTextJson` instead)
 * Change of default values: